created header viewholder and refactored previous view holders.
