+++ WordPress/src/main/java/org/wordpress/android/ui/posts/PrepublishingActionsAdapter.kt
@@ -8,25 +8,29 @@ import org.wordpress.android.WordPress
 import org.wordpress.android.ui.posts.PrepublishingActionItemUiState.PrepublishingActionUiState
 import org.wordpress.android.ui.posts.PrepublishingActionItemUiState.PrepublishingButtonUiState
 import org.wordpress.android.ui.posts.PrepublishingActionItemUiState.PrepublishingHomeHeaderUiState
+import org.wordpress.android.ui.posts.PrepublishingHomeViewHolder.PrepublishingHeaderListItemViewHolder
+import org.wordpress.android.ui.posts.PrepublishingHomeViewHolder.PrepublishingHomeListItemViewHolder
 import org.wordpress.android.ui.utils.UiHelpers
+import org.wordpress.android.util.image.ImageManager
 import javax.inject.Inject
 
 private const val headerViewType: Int = 1
 private const val actionItemViewType: Int = 2
 private const val publishButtonViewType: Int = 3
 
-class PrepublishingActionsAdapter(context: Context) : RecyclerView.Adapter<PrepublishingActionsListItemViewHolder>() {
+class PrepublishingActionsAdapter(context: Context) : RecyclerView.Adapter<PrepublishingHomeViewHolder>() {
     private var items: List<PrepublishingActionItemUiState> = listOf()
     @Inject lateinit var uiHelpers: UiHelpers
+    @Inject lateinit var imageManager: ImageManager
 
     init {
         (context.applicationContext as WordPress).component().inject(this)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PrepublishingActionsListItemViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PrepublishingHomeViewHolder {
         return when (viewType) {
-            headerViewType -> TODO()
-            actionItemViewType -> PrepublishingActionsListItemViewHolder(parent, uiHelpers)
+            headerViewType -> PrepublishingHeaderListItemViewHolder(parent, uiHelpers, imageManager)
+            actionItemViewType -> PrepublishingHomeListItemViewHolder(parent, uiHelpers)
             publishButtonViewType -> TODO()
             else -> throw NotImplementedError("Unknown ViewType")
         }
@@ -53,8 +57,8 @@ class PrepublishingActionsAdapter(context: Context) : RecyclerView.Adapter<Prepu
 
     override fun getItemCount(): Int = items.size
 
-    override fun onBindViewHolder(holder: PrepublishingActionsListItemViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: PrepublishingHomeViewHolder, position: Int) {
         val item = items[position]
-        holder.bind(item)
+        holder.onBind(item)
     }
 }
